Change fudge factor to 1e-2.
Add method to read from file and load the list of times from a
directory.

Change adaptive time step scheme. Now it looks at the net flux
for each cell not at each wall individually.

Add keyword to control the position of the inner boundary. It
doesn't have to necesarrily be at the semimajor axis.