Bug Fix:
If any of the boolean arrays `good` `badMin` or `badMax` evaluate
to False for all the entries, the code tries to pass an empty
array to log10Interp.ev() and it doesn't know what to do with it.
To avoid this situation I make sure the sum of the boolean arrays
is larger than 0 before trying to use them.